リファクタ: table 環境評価を責務別モジュールへ分割する
table.rs に同居していた任意引数パース・本体走査・セル構築のフリー関数を
table/{opts,body,cell}.rs へ切り出し、親には table() エントリと既存テストのみ残す。
振る舞いは変更しない。

Closes #146
