+++ crates/parser/src/evaluator/environment/table.rs
@@ -18,22 +18,18 @@
 //!   特殊属性が必要なセルだけ `\cell[span=N]{...}` にエスカレートする
 //! - `\caption{...}` — キャプション（出現位置が最初の行より前なら上、後なら下に配置）
 
-use document::{CaptionPosition, DocNode, InlineNode, TableCell, TableRow};
-use read_style::CounterName;
-use syntax::{
-  SyntaxKind,
-  ast::{CommandView, EnvironmentView},
-  green::GreenElement,
-  token::TokenKind,
-};
-use types::{ColumnAlign, ColumnWidth, Length};
+mod body;
+mod cell;
+mod opts;
 
-use crate::evaluator::{
-  EvalError, Evaluator,
-  environment::{body_scan, caption::extract_caption},
-  inline::{extract_inline_nodes, extract_inline_nodes_from_elements},
-  opt_args::{OptType, OptValue, collect_command_opt_args, collect_environment_opt_args},
-};
+use body::{resolve_column_count, scan_table_body};
+use document::DocNode;
+use opts::{collect_table_opts, parse_columns_spec, parse_widths_spec};
+use read_style::CounterName;
+use syntax::ast::EnvironmentView;
+use types::{ColumnAlign, ColumnWidth};
+
+use crate::evaluator::{EvalError, Evaluator};
 
 /// `table` 環境を評価する
 ///
@@ -93,476 +89,11 @@ pub(super) fn table(view: &EnvironmentView, evaluator: &mut Evaluator) -> Result
   }]);
 }
 
-/// `table` 環境の任意引数を集約した構造体
-struct TableOpts {
-  columns_spec: Option<String>,
-  widths_spec: Option<String>,
-  label: Option<String>,
-  breakable: bool,
-}
-
-/// `table` の任意引数（`columns` / `widths` / `label` / `breakable`）を収集してスカラー化する
-///
-/// 既定では `breakable` は `true`（改ページによる分割を許可）。
-fn collect_table_opts(view: &EnvironmentView) -> Result<TableOpts, EvalError> {
-  let opt_args = collect_environment_opt_args(
-    view,
-    &[
-      ("columns", OptType::String),
-      ("widths", OptType::String),
-      ("label", OptType::String),
-      ("breakable", OptType::Bool),
-    ],
-  )?;
-
-  let mut columns_spec: Option<String> = None;
-  let mut widths_spec: Option<String> = None;
-  let mut label: Option<String> = None;
-  let mut breakable = true;
-  for (key, value) in opt_args {
-    match (key.as_str(), value) {
-      ("columns", OptValue::String(s)) => columns_spec = Some(s),
-      ("widths", OptValue::String(s)) => widths_spec = Some(s),
-      ("label", OptValue::String(s)) => label = Some(s),
-      ("breakable", OptValue::Bool(b)) => breakable = b,
-      _ => {},
-    }
-  }
-
-  return Ok(TableOpts {
-    columns_spec,
-    widths_spec,
-    label,
-    breakable,
-  });
-}
-
-/// `columns="left center right"` の値を [`ColumnAlign`] の列に変換する
-fn parse_columns_spec(spec: &str, view: &EnvironmentView) -> Result<Vec<ColumnAlign>, EvalError> {
-  let invalid = || EvalError::InvalidOptArgValue {
-    name: "table".to_string(),
-    key: "columns".to_string(),
-    expected: "left / center / right の空白区切り".to_string(),
-    span: view.span().into(),
-  };
-  let tokens: Vec<&str> = spec.split_whitespace().collect();
-  if tokens.is_empty() {
-    return Err(invalid());
-  }
-  return tokens.iter().map(|t| ColumnAlign::from_keyword(t).ok_or_else(invalid)).collect();
-}
-
-/// `widths="auto auto 5cm 0.3 *"` の値を [`ColumnWidth`] の列に変換する
-fn parse_widths_spec(spec: &str, view: &EnvironmentView) -> Result<Vec<ColumnWidth>, EvalError> {
-  let invalid = || EvalError::InvalidOptArgValue {
-    name: "table".to_string(),
-    key: "widths".to_string(),
-    expected: "auto / <num>mm / <num>cm / 0〜1 の比率 / * の空白区切り".to_string(),
-    span: view.span().into(),
-  };
-  let tokens: Vec<&str> = spec.split_whitespace().collect();
-  if tokens.is_empty() {
-    return Err(invalid());
-  }
-  return tokens.iter().map(|t| parse_width_token(t).ok_or_else(invalid)).collect();
-}
-
-/// `widths=` の 1 トークンを [`ColumnWidth`] に変換する
-///
-/// 受理する形式: `auto` / `*` / `<num>mm` / `<num>cm`（サフィックスは大小無視）/
-/// `0` より大きく `1` 以下の小数（本文幅に対する比率）。
-fn parse_width_token(token: &str) -> Option<ColumnWidth> {
-  if token == "auto" {
-    return Some(ColumnWidth::Auto);
-  }
-  if token == "*" {
-    return Some(ColumnWidth::Flex);
-  }
-  let lower = token.to_ascii_lowercase();
-  if let Some(stripped) = lower.strip_suffix("mm") {
-    let value: f32 = stripped.parse().ok()?;
-    if !(value.is_finite() && value > 0.0) {
-      return None;
-    }
-    return Some(ColumnWidth::Fixed(Length::mm(value)));
-  }
-  if let Some(stripped) = lower.strip_suffix("cm") {
-    let value: f32 = stripped.parse().ok()?;
-    if !(value.is_finite() && value > 0.0) {
-      return None;
-    }
-    return Some(ColumnWidth::Fixed(Length::cm(value)));
-  }
-  let ratio: f32 = lower.parse().ok()?;
-  if !(ratio.is_finite() && ratio > 0.0 && ratio <= 1.0) {
-    return None;
-  }
-  return Some(ColumnWidth::Ratio(ratio));
-}
-
-/// 本体走査で収集した行・キャプション情報
-struct TableBody {
-  head: Vec<(TableRow, miette::SourceSpan)>,
-  rows: Vec<(TableRow, miette::SourceSpan)>,
-  caption: Option<Vec<InlineNode>>,
-  caption_position: CaptionPosition,
-}
-
-/// 本体から `\head` / `\row` / `\caption` を走査して [`TableBody`] に収集する
-///
-/// `\head` / `\caption` は高々 1 回（重複は [`EvalError::DuplicateCommandInEnvironment`]）。
-/// 行はソース位置とともに収集し、列数確定後のセル数検証に使う。キャプション位置は、
-/// `\caption` が最初の行（`\head` / `\row`）よりソース上で先に現れた場合のみ `Top`。
-fn scan_table_body(view: &EnvironmentView) -> Result<TableBody, EvalError> {
-  let source = view.source();
-  // 行はソース位置とともに収集し、列数確定後にセル数を検証する
-  let mut head: Vec<(TableRow, miette::SourceSpan)> = Vec::new();
-  let mut rows: Vec<(TableRow, miette::SourceSpan)> = Vec::new();
-  let mut caption: Option<Vec<InlineNode>> = None;
-  // `\caption` が最初の行（`\head` / `\row`）よりソース上で先に現れた場合のみ Top
-  let mut caption_position = CaptionPosition::Bottom;
-
-  if let Some(body) = view.body() {
-    for cmd_view in body_scan::strict_command_calls(
-      source,
-      body,
-      "table",
-      &["head", "row", "caption"],
-      "\\head と \\row と \\caption",
-    )? {
-      match cmd_view.name() {
-        "head" => {
-          if !head.is_empty() {
-            return Err(EvalError::DuplicateCommandInEnvironment {
-              env: "table".to_string(),
-              name: "head".to_string(),
-              span: cmd_view.span().into(),
-            });
-          }
-          head = extract_head(&cmd_view)?;
-        },
-        "row" => {
-          let span = cmd_view.span().into();
-          rows.push((extract_row(&cmd_view)?, span));
-        },
-        "caption" => {
-          if caption.is_some() {
-            return Err(EvalError::DuplicateCommandInEnvironment {
-              env: "table".to_string(),
-              name: "caption".to_string(),
-              span: cmd_view.span().into(),
-            });
-          }
-          if head.is_empty() && rows.is_empty() {
-            caption_position = CaptionPosition::Top;
-          }
-          caption = Some(extract_caption(&cmd_view)?);
-        },
-        _ => unreachable!("許可リスト外は strict_command_calls がエラーにする"),
-      }
-    }
-  }
-
-  return Ok(TableBody {
-    head,
-    rows,
-    caption,
-    caption_position,
-  });
-}
-
-/// `\head{\row{...} ...}` からヘッダ行を抽出する
-///
-/// 引数の直下に書けるのは `\row` のみ。トリビア（空白・改行・段落区切り・コメント・
-/// 引数の括弧）はスキップし、それ以外はエラーにする。
-fn extract_head(view: &CommandView) -> Result<Vec<(TableRow, miette::SourceSpan)>, EvalError> {
-  let _opt_args = collect_command_opt_args(view, &[])?;
-  let Some(arg) = view.first_arg() else {
-    return Err(EvalError::MissingCommandArgument {
-      name: "head".to_string(),
-      expected: "\\row コマンド".to_string(),
-      span: view.span().into(),
-    });
-  };
-  if view.args_count() > 1 {
-    return Err(EvalError::ExtraCommandArgument {
-      name: "head".to_string(),
-      span: view.span().into(),
-    });
-  }
-
-  let source = view.source();
-  let mut rows = Vec::new();
-  for child in arg.children {
-    match child {
-      GreenElement::Token(token) => match token.kind {
-        TokenKind::Whitespace
-        | TokenKind::Newline
-        | TokenKind::ParagraphBreak
-        | TokenKind::Comment
-        | TokenKind::LBrace
-        | TokenKind::RBrace => {},
-        _ => {
-          return Err(EvalError::UnexpectedContentInEnvironment {
-            env: "table".to_string(),
-            expected: "\\head の中の \\row".to_string(),
-            span: token.span.into(),
-          });
-        },
-      },
-      GreenElement::Node(node) => {
-        if node.kind == SyntaxKind::CommandCall {
-          let row_view = CommandView::new(node, source);
-          if row_view.name() == "row" {
-            let span = row_view.span().into();
-            rows.push((extract_row(&row_view)?, span));
-          } else {
-            return Err(EvalError::UnexpectedCommandInEnvironment {
-              env: "table".to_string(),
-              name: row_view.name().to_string(),
-              expected: "\\head の中の \\row".to_string(),
-              span: node.span.into(),
-            });
-          }
-        } else {
-          return Err(EvalError::UnexpectedContentInEnvironment {
-            env: "table".to_string(),
-            expected: "\\head の中の \\row".to_string(),
-            span: node.span.into(),
-          });
-        }
-      },
-    }
-  }
-  if rows.is_empty() {
-    return Err(EvalError::MissingCommandArgument {
-      name: "head".to_string(),
-      expected: "\\row コマンド".to_string(),
-      span: view.span().into(),
-    });
-  }
-  return Ok(rows);
-}
-
-/// `\row[rule_above]{A & B & \cell[span=2]{C}}` から 1 行を抽出する
-///
-/// 引数の子要素をトップレベルの `&`（[`TokenKind::Ampersand`]）で分割し、
-/// 各区画をセルに変換する。区画全体が `\cell` コマンドの場合は属性付きセルとして、
-/// それ以外はインライン内容のセルとして評価する。
-fn extract_row(view: &CommandView) -> Result<TableRow, EvalError> {
-  let opt_args = collect_command_opt_args(view, &[("rule_above", OptType::Bool)])?;
-  let rule_above = opt_args.iter().any(|(key, value)| key == "rule_above" && matches!(value, OptValue::Bool(true)));
-
-  let Some(arg) = view.first_arg() else {
-    return Err(EvalError::MissingCommandArgument {
-      name: "row".to_string(),
-      expected: "セル内容".to_string(),
-      span: view.span().into(),
-    });
-  };
-  if view.args_count() > 1 {
-    return Err(EvalError::ExtraCommandArgument {
-      name: "row".to_string(),
-      span: view.span().into(),
-    });
-  }
-
-  let source = view.source();
-  let mut cells: Vec<TableCell> = Vec::new();
-  let mut segment: Vec<GreenElement> = Vec::new();
-  for child in arg.children {
-    if let GreenElement::Token(token) = child
-      && token.kind == TokenKind::Ampersand
-    {
-      cells.push(build_cell(source, &segment)?);
-      segment.clear();
-    } else {
-      segment.push(*child);
-    }
-  }
-  cells.push(build_cell(source, &segment)?);
-
-  for cell in &cells {
-    if contains_line_break(&cell.content) {
-      return Err(EvalError::LineBreakInTableCell {
-        span: view.span().into(),
-      });
-    }
-  }
-
-  return Ok(TableRow { cells, rule_above });
-}
-
-/// セル内容に強制改行（`\\`）が含まれるかを再帰的に判定する
-fn contains_line_break(nodes: &[InlineNode]) -> bool {
-  return nodes.iter().any(|node| match node {
-    InlineNode::LineBreak => true,
-    InlineNode::Styled { children, .. } | InlineNode::Colored { children, .. } => contains_line_break(children),
-    _ => false,
-  });
-}
-
-/// `&` 分割後の 1 区画を [`TableCell`] に変換する
-///
-/// 区画の非トリビア要素が `\cell` コマンド 1 つだけなら属性付きセルとして評価し、
-/// `\cell` と他の内容が混在していればエラーにする。それ以外の区画は
-/// インライン内容として評価し、前後の空白をトリムする。
-fn build_cell(source: &str, elements: &[GreenElement]) -> Result<TableCell, EvalError> {
-  let mut cell_view: Option<CommandView> = None;
-  let mut has_other_content = false;
-  for element in elements {
-    match element {
-      GreenElement::Token(token) => match token.kind {
-        TokenKind::Whitespace | TokenKind::Newline | TokenKind::Comment | TokenKind::LBrace | TokenKind::RBrace => {},
-        _ => has_other_content = true,
-      },
-      GreenElement::Node(node) => {
-        if node.kind == SyntaxKind::CommandCall {
-          let candidate = CommandView::new(node, source);
-          if candidate.name() == "cell" {
-            if cell_view.is_some() {
-              // 同一区画に \cell が 2 つ — `&` の書き忘れ
-              return Err(EvalError::TableCellMixedContent {
-                span: node.span.into(),
-              });
-            }
-            cell_view = Some(candidate);
-          } else {
-            has_other_content = true;
-          }
-        } else {
-          has_other_content = true;
-        }
-      },
-    }
-  }
-
-  if let Some(cell_cmd) = cell_view {
-    if has_other_content {
-      return Err(EvalError::TableCellMixedContent {
-        span: cell_cmd.span().into(),
-      });
-    }
-    return extract_cell_command(&cell_cmd);
-  }
-
-  let content = extract_inline_nodes_from_elements(source, elements)?;
-  return Ok(TableCell::new(trim_cell_content(content)));
-}
-
-/// `\cell[span=N]{...}` を属性付きセルに変換する
-fn extract_cell_command(view: &CommandView) -> Result<TableCell, EvalError> {
-  let opt_args = collect_command_opt_args(view, &[("span", OptType::Number)])?;
-  let mut span: u32 = 1;
-  for (key, value) in opt_args {
-    if let ("span", OptValue::Number(n)) = (key.as_str(), value) {
-      // span は 1 以上の整数のみ受理する
-      if !(n.is_finite() && n >= 1.0 && n.fract() == 0.0 && n <= f64::from(u32::MAX)) {
-        return Err(EvalError::InvalidOptArgValue {
-          name: "cell".to_string(),
-          key: "span".to_string(),
-          expected: "1 以上の整数".to_string(),
-          span: view.span().into(),
-        });
-      }
-      #[allow(clippy::cast_sign_loss, clippy::cast_possible_truncation)]
-      {
-        span = n as u32;
-      }
-    }
-  }
-
-  let Some(arg) = view.first_arg() else {
-    return Err(EvalError::MissingCommandArgument {
-      name: "cell".to_string(),
-      expected: "セル内容".to_string(),
-      span: view.span().into(),
-    });
-  };
-  if view.args_count() > 1 {
-    return Err(EvalError::ExtraCommandArgument {
-      name: "cell".to_string(),
-      span: view.span().into(),
-    });
-  }
-
-  let content = trim_cell_content(extract_inline_nodes(view.source(), arg)?);
-  return Ok(TableCell { content, span });
-}
-
-/// セル内容の前後の空白由来 `Text` ノードをトリムする
-///
-/// `\row{Alice & 92}` の区切り前後の空白がセル内容に残ると揃えがずれるため、
-/// 先頭・末尾の空白のみの `Text` を除去し、境界の `Text` の端も削る。
-fn trim_cell_content(mut content: Vec<InlineNode>) -> Vec<InlineNode> {
-  while matches!(content.first(), Some(InlineNode::Text(t)) if t.trim().is_empty()) {
-    content.remove(0);
-  }
-  if let Some(InlineNode::Text(t)) = content.first_mut() {
-    *t = t.trim_start().to_string();
-  }
-  while matches!(content.last(), Some(InlineNode::Text(t)) if t.trim().is_empty()) {
-    content.pop();
-  }
-  if let Some(InlineNode::Text(t)) = content.last_mut() {
-    *t = t.trim_end().to_string();
-  }
-  return content;
-}
-
-/// 列数を決定し、全行のセル数（`span` 合計）が一致するか検証する
-///
-/// 列数は `columns` / `widths` の明示指定を優先し、両方未指定なら行のセル数（`span` 合計）の
-/// 最大値を採る。`columns` と `widths` の長さが食い違う場合は
-/// [`EvalError::TableColumnsWidthsMismatch`]、列数と一致しない行があれば
-/// [`EvalError::TableRowCellCountMismatch`] を返す。
-fn resolve_column_count(
-  columns_tokens: Option<&[ColumnAlign]>,
-  widths_tokens: Option<&[ColumnWidth]>,
-  head: &[(TableRow, miette::SourceSpan)],
-  rows: &[(TableRow, miette::SourceSpan)],
-  view: &EnvironmentView,
-) -> Result<usize, EvalError> {
-  // 列数の決定: columns / widths の明示指定を優先し、両方未指定なら行のセル数（span 合計）の最大値
-  let column_count = match (columns_tokens, widths_tokens) {
-    (Some(c), Some(w)) => {
-      if c.len() != w.len() {
-        return Err(EvalError::TableColumnsWidthsMismatch {
-          columns: c.len(),
-          widths: w.len(),
-          span: view.span().into(),
-        });
-      }
-      c.len()
-    },
-    (Some(c), None) => c.len(),
-    (None, Some(w)) => w.len(),
-    (None, None) => head.iter().chain(rows.iter()).map(|(row, _)| row_span_sum(row)).max().unwrap_or(0),
-  };
-
-  // 各行のセル数（span 合計）が列数と一致するか検証する
-  for (row, span) in head.iter().chain(rows.iter()) {
-    let actual = row_span_sum(row);
-    if actual != column_count {
-      return Err(EvalError::TableRowCellCountMismatch {
-        expected: column_count,
-        actual,
-        span: *span,
-      });
-    }
-  }
-
-  return Ok(column_count);
-}
-
-/// 行のセル数（`span` 合計）を返す
-fn row_span_sum(row: &TableRow) -> usize { return row.cells.iter().map(|cell| cell.span as usize).sum(); }
-
 #[cfg(test)]
 #[allow(clippy::unwrap_used)]
 mod tests {
   use bumpalo::Bump;
-  use document::inline_nodes_to_plain_text;
+  use document::{CaptionPosition, InlineNode, TableRow, inline_nodes_to_plain_text};
 
   use super::*;
   use crate::evaluator::lookup_env_parse_mode;
